+++ device_preview/lib/src/views/theme.dart
@@ -2,6 +2,7 @@ import 'package:device_preview/src/state/state.dart';
 import 'package:flutter/material.dart';
 
 extension ThemeBackgroundExtension on DevicePreviewBackgroundThemeData {
+  /// Converts a [DevicePreviewBackgroundThemeData] to a [ThemeData].
   ThemeData asThemeData() {
     switch (this) {
       case DevicePreviewBackgroundThemeData.dark:
@@ -13,6 +14,7 @@ extension ThemeBackgroundExtension on DevicePreviewBackgroundThemeData {
 }
 
 extension ThemeToolbarExtension on DevicePreviewToolBarThemeData {
+  /// Converts a [DevicePreviewToolBarThemeData] to a [ThemeData].
   ThemeData asThemeData() {
     switch (this) {
       case DevicePreviewToolBarThemeData.dark: